requirements.txt
boto3==1.38.4
pandas==2.2.3
sqlalchemy==2.0.40
python-dotenv==1.1.0
psycopg2-binary==2.9.10